+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] (1.1.0)
+### Added
+- New method `ReuseFileEntry.UpdateFileContents` to apply changes to REUSE metadata kept in a file.
+
 ## [1.0.0] - 2025-08-19
 ### Added
 - New classes: